Refactor lidar example DeepQ class into deepq.py
